Lint the codebook
