+++ src/styles/components/_codebook.scss
@@ -7,6 +7,7 @@
     font-size: .9em;
     color: var(--text-light);
     background-color: var(--architect-warning);
+    text-transform: lowercase;
   }
 
   &__category {
